BUG: No measurement frame was not taking into account output image orientation
When not using a measurement frame to specify the tensor orientation, but when instead using the image direction, the input image direction was taken into account, but the output image information was not taken into account. When resampling tensors in the same space then their original space, if the direction matrix of the output space was not the identity matrix, this was leading to the tensors being misaligned.