BUG: Remove unnecessary "project" command in CMakeLists.txt
The project command was called both in the main folder of the project, as well as in a subfolder. The second call was changing the value of ResampleDTIlogEuclidean_SOURCE_DIR and ResampleDTIlogEuclidean_BINARY_DIR which are used for the tests. The tests were failing on MacOS because the temporary folder for test results was created in a different folder than the one expected due to the modification of the variables value.